+++ packages/mobile/src/components/audio/Audio.tsx
@@ -39,11 +39,11 @@ import { apiClient } from 'app/services/audius-api-client'
 import { audiusBackendInstance } from 'app/services/audius-backend-instance'
 
 import { useChromecast } from './GoogleCast'
-import { logListen } from './listens'
 
 const { getUser } = cacheUsersSelectors
 const { getTrack } = cacheTracksSelectors
 const { getPlaying, getSeek, getCurrentTrack, getCounter } = playerSelectors
+const { recordListen } = tracksSocialActions
 const {
   getIndex,
   getOrder,
@@ -257,8 +257,10 @@ export const Audio = () => {
     }
   })
 
+  const trackId = track?.track_id
+
   const onProgress = useCallback(async () => {
-    if (!track || !currentUserId) return
+    if (!trackId || !currentUserId) return
     if (progressInvalidator.current) {
       progressInvalidator.current = false
       return
@@ -267,20 +269,11 @@ export const Audio = () => {
     const duration = await TrackPlayer.getDuration()
     const position = await TrackPlayer.getPosition()
 
-    // Replicates logic in dapp.
-    // TODO: REMOVE THIS ONCE BACKEND SUPPORTS THIS FEATURE
-    if (
-      position > RECORD_LISTEN_SECONDS &&
-      (track.owner_id !== currentUserId || track.play_count < 10) &&
-      !listenLoggedForTrack &&
-      // TODO: log listens for offline plays when reconnected
-      (!isOfflineModeEnabled || isReachable)
-    ) {
+    if (position > RECORD_LISTEN_SECONDS && !listenLoggedForTrack) {
       // Debounce logging a listen, update the state variable appropriately onSuccess and onFailure
       setListenLoggedForTrack(true)
-      logListen(track.track_id, currentUserId, () =>
-        setListenLoggedForTrack(false)
       )
+        dispatch(recordListen(trackId))
     }
     if (!isCasting) {
       // If we aren't casting, update the progress
@@ -296,7 +289,8 @@ export const Audio = () => {
     isOfflineModeEnabled,
     isReachable,
     listenLoggedForTrack,
-    track
+    trackId,
+    dispatch
   ])
 
   useEffect(() => {
@@ -331,11 +325,8 @@ export const Audio = () => {
   // Restart (counter) handler
   useEffect(() => {
     setSeekPosition(0)
-  }, [counter, setSeekPosition])
-
-  useEffect(() => {
     setListenLoggedForTrack(false)
-  }, [track, setListenLoggedForTrack])
+  }, [counter, setSeekPosition])
 
   const { loading: loadingOfflineTrack, value: offlineTrackUri } =
     useOfflineTrackUri(track?.track_id.toString())
